Moved VocabularyTable into a separate jsx component
